move stream to its own directory
